refactor api routes, make scraper ts safe

+++ pages/api/deleteRecipe/[...id].ts
@@ -4,6 +4,8 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth";
 import { z } from "zod";
 import { authOptions } from "../auth/[...nextauth]";
+import { Recipe } from "@/types/zod";
+import { DeleteObjectCommand, S3Client } from "@aws-sdk/client-s3";
 
 export default async function handler(
   req: NextApiRequest,
@@ -11,25 +13,36 @@ export default async function handler(
 ) {
   const session = await getServerSession(req, res, authOptions);
 
-  console.log(req.method);
+  if (!session || !session.user.id) return res.status(401);
+  if (req.method !== "DELETE") return res.status(405);
 
-  const client = await clientPromise;
-  const db = client.db("data");
+  const s3Client = new S3Client({});
+  const mongoClient = await clientPromise;
+  const db = mongoClient.db("data");
 
   const idCheck = z.string();
   const recipeId = idCheck.parse(req?.query?.id?.[0]);
 
   const o_id: Types.ObjectId = new mongoose.Types.ObjectId(recipeId);
 
-  if (!session) {
-    return res.status(400).json({ data: "no request" });
-  }
-
-  const post = await db.collection(`${session.user.id}`).findOneAndDelete({
+  const deletion = await db.collection(`${session.user.id}`).findOneAndDelete({
     _id: o_id,
   });
 
-  console.log(post);
+  if (deletion.ok && deletion.value && deletion.value.imageId) {
+    //if this post has an image, delete it
+    const command = new DeleteObjectCommand({
+      Bucket: "yrrb",
+      Key: `${session.user.id}/${deletion.value.imageId}`,
+    });
 
-  res.status(200).json(post);
+    try {
+      const response = await s3Client.send(command);
+      console.log(response);
+    } catch (error) {
+      console.error(error);
+    }
+  }
+
+  res.status(200);
 }
